+++ CHANGELOG.md
@@ -10,3 +10,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 - Option to use external MySQL database (to support legacy installations) - see [#48](https://github.com/ExpediaInc/apiary-metastore/issues/48).
+- Option to associate multiple VPCs to Service Discovery namespace - see
+[#66](https://github.com/ExpediaInc/apiary-metastore/issues/66)